remove legacy shops service
