+++ apps/cms/src/services/shops/index.ts
@@ -1,0 +1,4 @@
+export * from "./seoService";
+export * from "./themeService";
+export * from "./settingsService";
+export type { Shop, ShopSettings } from "@acme/types";
