+++ apps/cms/src/services/shops/settingsService.ts
@@ -1,0 +1,140 @@
+import type { ShopSettings } from "@acme/types";
+import { authorize, fetchSettings, persistSettings } from "./helpers";
+import {
+  parseCurrencyTaxForm,
+  parseDepositForm,
+  parseReverseLogisticsForm,
+  parseUpsReturnsForm,
+  parsePremierDeliveryForm,
+  parseAiCatalogForm,
+} from "./validation";
+
+export function getSettings(shop: string) {
+  return fetchSettings(shop);
+}
+
+export async function setFreezeTranslations(shop: string, freeze: boolean) {
+  await authorize();
+  const current = await fetchSettings(shop);
+  const updated: ShopSettings = { ...current, freezeTranslations: freeze };
+  await persistSettings(shop, updated);
+  return updated;
+}
+
+export async function updateCurrencyAndTax(
+  shop: string,
+  formData: FormData,
+): Promise<{ settings?: ShopSettings; errors?: Record<string, string[]> }> {
+  await authorize();
+  const { data, errors } = parseCurrencyTaxForm(formData);
+  if (!data) {
+    return { errors };
+  }
+  const current = await fetchSettings(shop);
+  const updated: ShopSettings = {
+    ...current,
+    currency: data.currency,
+    taxRegion: data.taxRegion,
+  };
+  await persistSettings(shop, updated);
+  return { settings: updated };
+}
+
+export async function updateDeposit(
+  shop: string,
+  formData: FormData,
+): Promise<{ settings?: ShopSettings; errors?: Record<string, string[]> }> {
+  await authorize();
+  const { data, errors } = parseDepositForm(formData);
+  if (!data) {
+    return { errors };
+  }
+  const current = await fetchSettings(shop);
+  const updated: ShopSettings = {
+    ...current,
+    depositService: {
+      enabled: data.enabled,
+      intervalMinutes: data.intervalMinutes,
+    },
+  };
+  await persistSettings(shop, updated);
+  return { settings: updated };
+}
+
+export async function updateReverseLogistics(
+  shop: string,
+  formData: FormData,
+): Promise<{ settings?: ShopSettings; errors?: Record<string, string[]> }> {
+  await authorize();
+  const { data, errors } = parseReverseLogisticsForm(formData);
+  if (!data) {
+    return { errors };
+  }
+  const current = await fetchSettings(shop);
+  const updated: ShopSettings = {
+    ...current,
+    reverseLogisticsService: {
+      enabled: data.enabled,
+      intervalMinutes: data.intervalMinutes,
+    },
+  };
+  await persistSettings(shop, updated);
+  return { settings: updated };
+}
+
+export async function updateUpsReturns(
+  shop: string,
+  formData: FormData,
+): Promise<{ settings?: ShopSettings; errors?: Record<string, string[]> }> {
+  await authorize();
+  const { data, errors } = parseUpsReturnsForm(formData);
+  if (!data) {
+    return { errors };
+  }
+  const current = await fetchSettings(shop);
+  const updated: ShopSettings = {
+    ...current,
+    returnService: { upsEnabled: data.enabled },
+  };
+  await persistSettings(shop, updated);
+  return { settings: updated };
+}
+
+export async function updatePremierDelivery(
+  shop: string,
+  formData: FormData,
+): Promise<{ settings?: ShopSettings; errors?: Record<string, string[]> }> {
+  await authorize();
+  const { data, errors } = parsePremierDeliveryForm(formData);
+  if (!data) {
+    return { errors };
+  }
+  const current = await fetchSettings(shop);
+  const updated: ShopSettings = {
+    ...current,
+    premierDelivery: data,
+  };
+  await persistSettings(shop, updated);
+  return { settings: updated };
+}
+
+export async function updateAiCatalog(
+  shop: string,
+  formData: FormData,
+): Promise<{ settings?: ShopSettings; errors?: Record<string, string[]> }> {
+  await authorize();
+  const { data, errors } = parseAiCatalogForm(formData);
+  if (!data) {
+    return { errors };
+  }
+  const current = await fetchSettings(shop);
+  const seo = { ...(current.seo ?? {}) };
+  (seo as any).aiCatalog = {
+    enabled: data.enabled,
+    fields: data.fields,
+    pageSize: data.pageSize,
+  };
+  const updated: ShopSettings = { ...current, seo };
+  await persistSettings(shop, updated);
+  return { settings: updated };
+}
